modularise bdk project code   * delete bdk.rs file
