+++ src/wallet/bdk_wallet/mod.rs
@@ -1,0 +1,4 @@
+mod keychain;
+mod wallet;
+
+pub use keychain::Keychain;
